requirements.txt
parameterized==0.9.0
viam-sdk==0.2.18
